Successfully plugged wynton into apprising, just tested it!
Nothing shows up at /blog/ because the templates are not defined at all, but that is a job for another day. Next step: finish testing by defining a simple template just to make sure everything is working. Then working on plugging wes in. Wes also needs to have the admin page written.